+++ src/utils/index.ts
@@ -1,8 +1,6 @@
-export {buildCommandArgs} from './buildCommandArgs';
-export {humanReadableArgName} from './humanReadableArgName';
-export {matchCommand} from './matchCommand';
-export {pad} from './pad';
-export {padRow} from './padRow';
-export {parseArgs} from './parseArgs';
-export {parseCommand} from './parseCommand';
-export {prettifyArray} from './prettifyArray';
+export * from './formatting';
+export * from './buildCommandArgs';
+export * from './humanReadableArgName';
+export * from './matchCommand';
+export * from './parseArgs';
+export * from './parseCommand';
