--- src/utils/padRow.ts
@@ -1,11 +1,0 @@
-import {PADDING} from '../constants';
-
-/**
- * Pad a row on the start and end with spaces.
- */
-export function padRow(value: string): string {
-  return value
-    .split('\n')
-    .map(row => PADDING + row + PADDING)
-    .join('\n');
-}
